@@ -2,8 +2,8 @@ Tony612Com::Application.routes.draw do
   devise_for :admins
 
   resources :posts
-  match "about" => "admins#show"
+  get "about" => "admins#show"
   root to: "posts#index"
-  match "life" => "posts#life"
-  match "tech" => "posts#tech"
+  get "life" => "posts#life"
+  get "tech" => "posts#tech"
 end
